+++ controllers/operator/apigateway_controller_test.go
@@ -2,12 +2,12 @@ package operator
 
 import (
 	"context"
+	"time"
 
 	"github.com/go-logr/logr"
+	gatewayv1beta1 "github.com/kyma-project/api-gateway/apis/gateway/v1beta1"
 	operatorv1alpha1 "github.com/kyma-project/api-gateway/apis/operator/v1alpha1"
-	"github.com/kyma-project/api-gateway/controllers"
-	"github.com/kyma-project/api-gateway/internal/reconciliations/custom_resource"
-	"github.com/pkg/errors"
+	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -21,15 +21,11 @@ var _ = Describe("API-Gateway Controller", func() {
 	Context("Reconcile", func() {
 		It("Should not return an error when CR was not found", func() {
 			// given
-			apiClient := createFakeClient()
-
+			c := createFakeClient()
 			agr := &APIGatewayReconciler{
-				Client: apiClient,
+				Client: c,
 				Scheme: getTestScheme(),
-				apiGatewayReconciliation: &apiGatewayReconciliationMock{
-					status: controllers.ReadyStatus(),
-				},
-				log: logr.Discard(),
+				log:    logr.Discard(),
 			}
 
 			// when
@@ -40,55 +36,20 @@ var _ = Describe("API-Gateway Controller", func() {
 			Expect(result).Should(Equal(reconcile.Result{}))
 		})
 
-		It("Should not requeue a CR without finalizers, because it's considered to be in deletion", func() {
-			// given
-			istioCR := &operatorv1alpha1.APIGateway{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      apiGatewayCRName,
-					Namespace: testNamespace,
-				},
-			}
-
-			apiClient := createFakeClient(istioCR)
-
-			agr := &APIGatewayReconciler{
-				Client: apiClient,
-				Scheme: getTestScheme(),
-				apiGatewayReconciliation: &apiGatewayReconciliationMock{
-					status: controllers.ReadyStatus(),
-				},
-				log: logr.Discard(),
-			}
-
-			// when
-			result, err := agr.Reconcile(context.Background(), reconcile.Request{NamespacedName: types.NamespacedName{Namespace: testNamespace, Name: apiGatewayCRName}})
-
-			// then
-			Expect(err).ShouldNot(HaveOccurred())
-			Expect(result).Should(Equal(reconcile.Result{}))
-		})
-
-		It("Should set status to ready when API-Gateway reconciliation succeed", func() {
+		It("Should add finalizer when API-Gateway CR is not marked for deletion", func() {
 			// given
 			apiGatewayCR := &operatorv1alpha1.APIGateway{
 				ObjectMeta: metav1.ObjectMeta{
 					Name:      apiGatewayCRName,
 					Namespace: testNamespace,
-					Finalizers: []string{
-						custom_resource.ApiGatewayFinalizer,
-					},
 				},
 			}
 
-			apiClient := createFakeClient(apiGatewayCR)
-
+			c := createFakeClient(apiGatewayCR)
 			agr := &APIGatewayReconciler{
-				Client: apiClient,
+				Client: c,
 				Scheme: getTestScheme(),
-				apiGatewayReconciliation: &apiGatewayReconciliationMock{
-					status: controllers.ReadyStatus(),
-				},
-				log: logr.Discard(),
+				log:    logr.Discard(),
 			}
 
 			// when
@@ -98,31 +59,88 @@ var _ = Describe("API-Gateway Controller", func() {
 			Expect(err).ShouldNot(HaveOccurred())
 			Expect(result).Should(Equal(reconcile.Result{}))
 
-			Expect(apiClient.Get(context.TODO(), client.ObjectKeyFromObject(apiGatewayCR), apiGatewayCR)).Should(Succeed())
+			Expect(c.Get(context.TODO(), client.ObjectKeyFromObject(apiGatewayCR), apiGatewayCR)).Should(Succeed())
+			Expect(apiGatewayCR.GetObjectMeta().GetFinalizers()).To(ContainElement(ApiGatewayFinalizer))
+		})
+
+		It("Should set status to Ready when API-Gateway reconciliation succeed", func() {
+			// given
+			apiGatewayCR := &operatorv1alpha1.APIGateway{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:       apiGatewayCRName,
+					Namespace:  testNamespace,
+					Finalizers: []string{ApiGatewayFinalizer},
+				},
+			}
+
+			c := createFakeClient(apiGatewayCR)
+			agr := &APIGatewayReconciler{
+				Client: c,
+				Scheme: getTestScheme(),
+				log:    logr.Discard(),
+			}
+
+			// when
+			result, err := agr.Reconcile(context.Background(), reconcile.Request{NamespacedName: types.NamespacedName{Namespace: testNamespace, Name: apiGatewayCRName}})
+
+			// then
+			Expect(err).ShouldNot(HaveOccurred())
+			Expect(result).Should(Equal(reconcile.Result{}))
+
+			Expect(c.Get(context.TODO(), client.ObjectKeyFromObject(apiGatewayCR), apiGatewayCR)).Should(Succeed())
 			Expect(apiGatewayCR.Status.State).Should(Equal(operatorv1alpha1.Ready))
 		})
 
-		It("Should set error status and return an error when API-Gateway reconciliation failed", func() {
+		It("Should delete API-Gateway CR if there are no blocking resources", func() {
 			// given
-			apiGatewayCR := &operatorv1alpha1.APIGateway{
+			now := metav1.NewTime(time.Now())
+			apiGatewayCR := &operatorv1alpha1.APIGateway{ObjectMeta: metav1.ObjectMeta{
+				Name:              apiGatewayCRName,
+				Namespace:         testNamespace,
+				DeletionTimestamp: &now,
+				Finalizers:        []string{ApiGatewayFinalizer},
+			},
+			}
+
+			c := createFakeClient(apiGatewayCR)
+			agr := &APIGatewayReconciler{
+				Client: c,
+				Scheme: getTestScheme(),
+				log:    logr.Discard(),
+			}
+
+			// when
+			result, err := agr.Reconcile(context.Background(), reconcile.Request{NamespacedName: types.NamespacedName{Namespace: testNamespace, Name: apiGatewayCRName}})
+
+			// then
+			Expect(err).ShouldNot(HaveOccurred())
+			Expect(result).Should(Equal(reconcile.Result{}))
+
+			Expect(errors.IsNotFound(c.Get(context.TODO(), client.ObjectKeyFromObject(apiGatewayCR), apiGatewayCR))).To(BeTrue())
+		})
+
+		It("Should not delete API-Gateway CR if there is any APIRule on cluster", func() {
+			// given
+			now := metav1.NewTime(time.Now())
+			apiGatewayCR := &operatorv1alpha1.APIGateway{ObjectMeta: metav1.ObjectMeta{
+				Name:              apiGatewayCRName,
+				Namespace:         testNamespace,
+				DeletionTimestamp: &now,
+				Finalizers:        []string{ApiGatewayFinalizer},
+			},
+			}
+			apiRule := &gatewayv1beta1.APIRule{
 				ObjectMeta: metav1.ObjectMeta{
-					Name:      apiGatewayCRName,
-					Namespace: testNamespace,
-					Finalizers: []string{
-						custom_resource.ApiGatewayFinalizer,
-					},
+					Name:      "api-rule",
+					Namespace: "default",
 				},
 			}
 
-			apiClient := createFakeClient(apiGatewayCR)
-
+			c := createFakeClient(apiGatewayCR, apiRule)
 			agr := &APIGatewayReconciler{
-				Client: apiClient,
+				Client: c,
 				Scheme: getTestScheme(),
-				apiGatewayReconciliation: &apiGatewayReconciliationMock{
-					status: controllers.ErrorStatus(errors.New("API-Gateway CR reconciliation error"), "Test error description"),
-				},
-				log: logr.Discard(),
+				log:    logr.Discard(),
 			}
 
 			// when
@@ -130,20 +148,13 @@ var _ = Describe("API-Gateway Controller", func() {
 
 			// then
 			Expect(err).Should(HaveOccurred())
-			Expect(err.Error()).To(ContainSubstring("API-Gateway CR reconciliation error"))
+			Expect(err.Error()).To(Equal("could not delete API-Gateway CR since there are 1 APIRule(s) present that block its deletion"))
 			Expect(result).Should(Equal(reconcile.Result{}))
 
-			Expect(apiClient.Get(context.TODO(), client.ObjectKeyFromObject(apiGatewayCR), apiGatewayCR)).Should(Succeed())
-			Expect(apiGatewayCR.Status.State).Should(Equal(operatorv1alpha1.Error))
-			Expect(apiGatewayCR.Status.Description).To(ContainSubstring("Test error description"))
+			Expect(c.Get(context.TODO(), client.ObjectKeyFromObject(apiGatewayCR), apiGatewayCR)).Should(Succeed())
+			Expect(apiGatewayCR.Status.State).To(Equal(operatorv1alpha1.Warning))
+			Expect(apiGatewayCR.Status.Description).To(Equal("There are APIRule(s) that block the deletion of API-Gateway CR. Please take a look at kyma-system/api-gateway-controller-manager logs to see more information about the warning"))
+			Expect(apiGatewayCR.GetObjectMeta().GetFinalizers()).To(ContainElement(ApiGatewayFinalizer))
 		})
 	})
 })
-
-type apiGatewayReconciliationMock struct {
-	status controllers.Status
-}
-
-func (i *apiGatewayReconciliationMock) Reconcile(_ context.Context, apiGatewayCR *operatorv1alpha1.APIGateway) controllers.Status {
-	return i.status
-}
