+++ controllers/operator/apigateway_types.go
@@ -2,7 +2,6 @@ package operator
 
 import (
 	"github.com/go-logr/logr"
-	"github.com/kyma-project/api-gateway/internal/reconciliations/custom_resource"
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -10,7 +9,6 @@ import (
 // APIGatewayReconciler reconciles a APIGateway object
 type APIGatewayReconciler struct {
 	client.Client
-	Scheme                   *runtime.Scheme
-	log                      logr.Logger
-	apiGatewayReconciliation custom_resource.ApiGatewayReconciliation
+	Scheme *runtime.Scheme
+	log    logr.Logger
 }
