--- internal/reconciliations/custom_resource/custom_resource_suite_test.go
@@ -1,34 +1,0 @@
-package custom_resource
-
-import (
-	"testing"
-
-	gatewayv1beta1 "github.com/kyma-project/api-gateway/apis/gateway/v1beta1"
-	operatorv1alpha1 "github.com/kyma-project/api-gateway/apis/operator/v1alpha1"
-	"github.com/kyma-project/api-gateway/tests"
-	. "github.com/onsi/ginkgo/v2"
-	"github.com/onsi/ginkgo/v2/types"
-	. "github.com/onsi/gomega"
-	"k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset/scheme"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
-)
-
-func TestAPIs(t *testing.T) {
-	RegisterFailHandler(Fail)
-
-	RunSpecs(t, "API-Gateway Operator Reconciliation Suite")
-}
-
-var _ = ReportAfterSuite("custom reporter", func(report types.Report) {
-	tests.GenerateGinkgoJunitReport("api-gateway-operator-reconciliation-suite", report)
-})
-
-func createFakeClient(objects ...client.Object) client.Client {
-	err := operatorv1alpha1.AddToScheme(scheme.Scheme)
-	Expect(err).ShouldNot(HaveOccurred())
-	err = gatewayv1beta1.AddToScheme(scheme.Scheme)
-	Expect(err).ShouldNot(HaveOccurred())
-
-	return fake.NewClientBuilder().WithScheme(scheme.Scheme).WithObjects(objects...).WithStatusSubresource(objects...).Build()
-}
