+++ tests/integration/testsuites/gateway/features/kyma_gateway.feature
@@ -31,14 +31,6 @@ Feature: Checking default kyma gateway
     And there "is no" "ClusterRoleBinding" "oathkeeper-maester-role-binding" in the cluster
     And there "is no" "PeerAuthentication" "ory-oathkeeper-maester-metrics" in namespace "kyma-system"
 
-  Scenario: Kyma Gateway is not removed when there is an APIRule
-    Given there is an "kyma-rule" APIRule with Gateway "kyma-system/kyma-gateway"
-    When disabling Kyma gateway
-    Then APIGateway CR is in "Warning" state with description "There are custom resources that block the deletion of Kyma Gateway. Please take a look at kyma-system/api-gateway-controller-manager logs to see more information about the warning"
-    And there is Istio Gateway "kyma-gateway" in "kyma-system" namespace
-    And APIRule "kyma-rule" is removed
-    And APIGateway CR is in "Ready" state with description ""
-
   Scenario: Kyma Gateway is not removed when there is a VirtualService
     Given there is an "kyma-vs" VirtualService with Gateway "kyma-system/kyma-gateway"
     When disabling Kyma gateway
